# 🚀 Story navigation performance improvements.