--- a/extensions/amp-story/1.0/amp-story.js
+++ b/extensions/amp-story/1.0/amp-story.js
@@ -894,7 +894,6 @@ export class AmpStory extends AMP.BaseElement {
     const steps = [
       () => {
         oldPage && oldPage.element.removeAttribute('active');
-        oldPage && oldPage.stopListeningToNavigationEvents();
 
         this.systemLayer_.setActivePageId(targetPageId);