--- a/extensions/amp-story/1.0/amp-story.js
+++ b/extensions/amp-story/1.0/amp-story.js
@@ -882,6 +882,88 @@ export class AmpStory extends AMP.BaseElement {
    * @return {!Promise}
    */
   switchTo_(targetPageId) {
+    if (isExperimentOn(this.win, 'amp-story-navigation-performance')) {
+      return this.experimentalSwitchTo_(targetPageId);
+    }
+
+    const targetPage = this.getPageById(targetPageId);
+    const pageIndex = this.getPageIndex(targetPage);
+
+    this.storeService_.dispatch(Action.CHANGE_PAGE, {
+      id: targetPageId,
+      index: pageIndex,
+    });
+
+    this.handlePreviewAttributes_(targetPage);
+
+    this.updateBackground_(targetPage.element, /* initial */ !this.activePage_);
+
+    if (targetPage.isAd()) {
+      this.storeService_.dispatch(Action.TOGGLE_AD, true);
+      setAttributeInMutate(this, Attributes.AD_SHOWING);
+    } else {
+      this.storeService_.dispatch(Action.TOGGLE_AD, false);
+      removeAttributeInMutate(this, Attributes.AD_SHOWING);
+      // TODO(alanorozco): decouple this using NavigationState
+      this.systemLayer_.setActivePageId(targetPageId);
+    }
+
+    // TODO(alanorozco): check if autoplay
+    this.navigationState_.updateActivePage(
+        pageIndex,
+        this.getPageCount(),
+        targetPage.element.id,
+        targetPage.getNextPageId() === null /* isFinalPage */
+    );
+
+    const oldPage = this.activePage_;
+
+    this.activePage_ = targetPage;
+
+    this.systemLayer_.resetDeveloperLogs();
+    this.systemLayer_.setDeveloperLogContextString(this.activePage_.element.id);
+
+    return targetPage.beforeVisible().then(() => {
+      this.triggerActiveEventForPage_();
+
+      if (oldPage) {
+        oldPage.setState(PageState.NOT_ACTIVE);
+
+        // Indication that this should be offscreen to left in desktop view.
+        setAttributeInMutate(oldPage, Attributes.VISITED);
+      }
+
+      // Starts playing the page, if the story is not paused.
+      // Note: navigation is prevented when the story is paused, this test
+      // covers the case where the story is rendered paused (eg: consent).
+      if (!this.storeService_.get(StateProperty.PAUSED_STATE)) {
+        targetPage.setState(PageState.ACTIVE);
+      }
+
+      // If first navigation.
+      if (!oldPage) {
+        this.registerAndPreloadBackgroundAudio_();
+      }
+
+      if (!this.storeService_.get(StateProperty.MUTED_STATE)) {
+        oldPage && oldPage.muteAllMedia();
+        this.activePage_.unmuteAllMedia();
+      }
+
+      this.preloadPagesByDistance_();
+      this.forceRepaintForSafari_();
+      this.maybePreloadBookend_();
+    });
+  }
+
+
+  /**
+   * Switches to a particular page.
+   * Protected behing the 'amp-story-navigation-performance' experiment.
+   * @param {string} targetPageId
+   * @return {!Promise}
+   */
+  experimentalSwitchTo_(targetPageId) {
     const targetPage = this.getPageById(targetPageId);
     const pageIndex = this.getPageIndex(targetPage);