Organizes the page navigation method in three steps.
Each step runs in its own browser frame: we execute the first step, and wait for the next frame before executing the second step, etc.

The time measured is the time needed for the next page to be displayed, with its media elements playing.

| Device  | Avg navigation time before | Avg navigation time after | % improvement |
| :-------------: | :-------------: | :-------------: | :-------------: |
| BLU 4.0  | 1742.71ms | 832.15ms | +109.42% |
| Nexus 5  | 272.14ms | 89.28ms | +204.84% |
| iPhone 5  | 580.38ms | 279.12ms | +107.93% |
| Chrome mobile emulation  | 77.36ms | 29.42ms | +107.93% |
| **Average**  | **-** | **-** | **+146.29%** |

I've been testing a lot of stories with media / no media / auto-advance / etc, on both mobile and desktop versions, hopefully this comes bug free :)))