@newmuis 2018-07-19
re: making sure we don't break it, what do you think about putting the new `switchTo` behind an experiment.  We can have:

```javascript
switchTo_(...) {
  if (isExperimentOn('amp-story-nav-improvements')) {
    this.newSwitchTo_(...);
  } else {
    this.oldSwitchTo_(...);
  }
}
```

Then we can gradually ramp up the experiment from 0% to 100%.

As a bonus, if we can figure out how to log the metrics, we can get a full view of how much this improves things for people, since we can A/B test (though it may be hard to control for device, browser, and page content).