@gmajoulet 2018-07-23
The feature now lives behind a new `amp-story-navigation-performance` experiment :)

PTAL